pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-video"
version = "0.1.0"
description = "AI-powered video analysis and prompt generation for video recreation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "opencv-python>=4.9.0.80",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ai-video = "ai_video.cli:app"

[tool.setuptools]
packages = ["ai_video"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
